# [Topology.Container.Grid] trivial optimization in GridTopology