Just call getValue once and use the const ref afterwards.

This is particularly showing for the function getIndex() which is called a lot of times (really a lot)

Release mode (full optim), MSVC and scene from Components/topology/RegularGridTopology.scn:
before:  `5000 iterations done in 22.5747 s ( 221.487 FPS).`
after  : `5000 iterations done in 20.7355 s ( 241.132 FPS).`
so 10% speedup

But the optimization is more  effective the more complex  the mesh is
(I changed the mesh in the previous scene with a liver with 500K triangles and the speed up was like around 30%)


In any case, it shows that `getValue()` is not really "free" and must be called only once if the value is needed more than one time, especially in the "hot" functions.

______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).